src/main/java/ooga/controller/CSVParser.java
package ooga.controller;

import com.opencsv.CSVReader;
import com.opencsv.exceptions.CsvValidationException;
import ooga.view.MapWrapper;

import java.io.FileReader;
import java.io.IOException;
import java.util.Arrays;

/**
 * This class is responsible for parsing the CSV file and creating the map.
 * @author Nick Ward
 */
public class CSVParser {
  private FileReader myFileReader;

  /**
   * Parses the data from the csv file and returns the map wrapper
   * @param csvPath
   * @return
   */
  public MapWrapper parseData(String csvPath) throws IllegalStateException {
    try {
      myFileReader = new FileReader(csvPath);
      MapWrapper mapWrapper = new MapWrapper();
      CSVReader csvReader = new CSVReader(myFileReader);
      String[] nextRecord;
      int row = 0;
      while ((nextRecord = csvReader.readNext()) != null) {
        // if any strings in nextRecord start with "#" or "", ignore them
        nextRecord = Arrays.stream(nextRecord).filter(s -> !s.startsWith("#") && !s.isEmpty()).toArray(String[]::new);
        if (nextRecord.length > 0) {
          if (row >= 0) {
            mapWrapper.addRow();
            for (String cell : nextRecord) {
              mapWrapper.addValueToRow(row, Integer.parseInt(cell.trim()));
            }
          }
          row++;
        }
      }
      return mapWrapper;
    } catch (CsvValidationException | IOException e) {
      throw new IllegalStateException("badCsvFile", e);
    }
  }
}
